requirements.txt
bittensor
substrate-interface
python-dotenv
bcrypt
bittensor-cli